[ITensors] `HDF5.jl` package extension (#1382)
