+++ ITensorGaussianMPS/Project.toml
@@ -1,7 +1,7 @@
 name = "ITensorGaussianMPS"
 uuid = "2be41995-7c9f-4653-b682-bfa4e7cebb93"
 authors = ["Matthew Fishman <mfishman@flatironinstitute.org> and contributors"]
-version = "0.1.4"
+version = "0.1.5"
 
 [deps]
 Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
@@ -10,6 +10,6 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [compat]
 Compat = "3.40.0, 4"
-ITensors = "0.3.58"
+ITensors = "0.3.58, 0.4"
 LinearAlgebra = "1.6"
 julia = "1.6"
