+++ ITensorVisualizationBase/Project.toml
@@ -1,7 +1,7 @@
 name = "ITensorVisualizationBase"
 uuid = "cd2553d2-8bef-4d93-8a38-c62f17d5ad23"
 authors = ["Matthew Fishman <mfishman@flatironinstitute.org> and contributors"]
-version = "0.1.5"
+version = "0.1.6"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -20,7 +20,7 @@ AbstractTrees = "0.4"
 Compat = "3.40.0, 4"
 GeometryBasics = "0.4.1"
 Graphs = "1.4.1"
-ITensors = "0.2.12, 0.3"
+ITensors = "0.2.12, 0.3, 0.4"
 MetaGraphs = "0.7.1"
 NetworkLayout = "0.4.3"
 Statistics = "1"
