+++ docs/src/examples/MPSandMPO.md
@@ -384,6 +384,9 @@ can be brought into orthogonal form by calling `psi = orthogonalize(psi, 1)`.
 
 ## Write and Read an MPS or MPO to Disk with HDF5
 
+!!! info 
+    Make sure to install the HDF5 package to use this feature. (Run `julia> ] add HDF5` in the Julia REPL console.)
+
 **Writing an MPS to an HDF5 File**
 
 Let's say you have an MPS `psi` which you have made or obtained
@@ -391,7 +394,7 @@ from a calculation. To write it to an HDF5 file named "myfile.h5"
 you can use the following pattern:
 
 ```julia
-using ITensors.HDF5
+using HDF5
 f = h5open("myfile.h5","w")
 write(f,"psi",psi)
 close(f)
@@ -402,8 +405,6 @@ or "Result MPS" and doesn't have to have the same name as the reference `psi`.
 Closing the file `f` is optional and you can also write other objects to the same
 file before closing it.
 
-[*Above we did `using ITensors.HDF5` since HDF5 is already included as a dependency with ITensor. You can also do `using HDF5` but must add the HDF5 package beforehand for that to work.*]
-
 **Reading an MPS from an HDF5 File**
 
 Say you have an HDF5 file "myfile.h5" which contains an MPS stored as a dataset with the
@@ -411,7 +412,7 @@ name "psi". (Which would be the situation if you wrote it as in the example abov
 To read this ITensor back from the HDF5 file, use the following pattern:
 
 ```julia
-using ITensors.HDF5
+using HDF5
 f = h5open("myfile.h5","r")
 psi = read(f,"psi",MPS)
 close(f)
@@ -435,7 +436,6 @@ H = MPO(os,sites)
 # Compute <psi|H|psi>
 energy_psi = inner(psi',H,psi)
 ```
-
 
 Note the `MPS` argument to the read function, which tells Julia which read function
 to call and how to interpret the data stored in the HDF5 dataset named "psi". In the
