+++ src/qn/qn.jl
@@ -399,27 +399,3 @@ function show(io::IO, q::QN)
   end
   return print(io, ")")
 end
-
-function write(parent::Union{HDF5.File,HDF5.Group}, gname::AbstractString, q::QN)
-  g = create_group(parent, gname)
-  attributes(g)["type"] = "QN"
-  attributes(g)["version"] = 1
-  names = [String(name(q[n])) for n in 1:maxQNs]
-  vals = [val(q[n]) for n in 1:maxQNs]
-  mods = [modulus(q[n]) for n in 1:maxQNs]
-  write(g, "names", names)
-  write(g, "vals", vals)
-  return write(g, "mods", mods)
-end
-
-function read(parent::Union{HDF5.File,HDF5.Group}, name::AbstractString, ::Type{QN})
-  g = open_group(parent, name)
-  if read(attributes(g)["type"]) != "QN"
-    error("HDF5 group or file does not contain QN data")
-  end
-  names = read(g, "names")
-  vals = read(g, "vals")
-  mods = read(g, "mods")
-  mqn = ntuple(n -> QNVal(names[n], vals[n], mods[n]), maxQNs)
-  return QN(mqn)
-end
